+++ pkgs/native_assets_builder/test/build_runner/concurrency_test_helper.dart
@@ -28,26 +28,26 @@ void main(List<String> args) async {
     singleHookTimeout: timeout,
     fileSystem: const LocalFileSystem(),
   ).build(
-    configCreator: () => BuildConfigBuilder()
-      ..setupCodeConfig(
+    inputCreator: () => BuildInputBuilder()
+      ..config.setupCode(
         targetArchitecture: Architecture.current,
         targetOS: targetOS,
         linkModePreference: LinkModePreference.dynamic,
-        cCompilerConfig: dartCICompilerConfig,
-        macOSConfig: targetOS == OS.macOS
+        cCompiler: dartCICompilerConfig,
+        macOS: targetOS == OS.macOS
             ? MacOSConfig(targetVersion: defaultMacOSVersion)
             : null,
       ),
     workingDirectory: packageUri,
     linkingEnabled: false,
     buildAssetTypes: [CodeAsset.type, DataAsset.type],
-    configValidator: (config) async => [
-      ...await validateDataAssetBuildConfig(config),
-      ...await validateCodeAssetBuildConfig(config),
+    inputValidator: (input) async => [
+      ...await validateDataAssetBuildInput(input),
+      ...await validateCodeAssetBuildInput(input),
     ],
-    buildValidator: (config, output) async => [
-      ...await validateCodeAssetBuildOutput(config, output),
-      ...await validateDataAssetBuildOutput(config, output),
+    buildValidator: (input, output) async => [
+      ...await validateCodeAssetBuildOutput(input, output),
+      ...await validateDataAssetBuildOutput(input, output),
     ],
     applicationAssetValidator: validateCodeAssetInApplication,
   );