+++ pkgs/native_assets_builder/test/build_runner/resources_test.dart
@@ -35,7 +35,7 @@ void main() async {
           dartExecutable,
           linkingEnabled: true,
           buildAssetTypes: [DataAsset.type],
-          configValidator: validateDataAssetBuildConfig,
+          inputValidator: validateDataAssetBuildInput,
           buildValidator: validateDataAssetBuildOutput,
           applicationAssetValidator: (_) async => [],
         ))!;
@@ -54,7 +54,7 @@ void main() async {
           buildResult: buildResult,
           resourceIdentifiers: resourcesUri,
           buildAssetTypes: [DataAsset.type],
-          configValidator: validateDataAssetLinkConfig,
+          inputValidator: validateDataAssetLinkInput,
           linkValidator: validateDataAssetLinkOutput,
           applicationAssetValidator: (_) async => [],
         );