+++ pkgs/swift2objc/lib/src/ast/_core/shared/parameter.dart
@@ -2,10 +2,11 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+import '../../ast_node.dart';
 import 'referred_type.dart';
 
 /// Describes parameters of function-like entities (e.g methods).
-class Parameter {
+class Parameter extends AstNode {
   String name;
   String? internalName;
   ReferredType type;
@@ -18,4 +19,10 @@ class Parameter {
 
   @override
   String toString() => '$name $internalName: $type';
+
+  @override
+  void visitChildren(Visitor visitor) {
+    super.visitChildren(visitor);
+    visitor.visit(type);
+  }
 }